+++ miniprogram/config/index.js
@@ -1,5 +1,5 @@
 const ENV = "production";
-const VERSION = "1.1.0";
+const VERSION = "1.1.1";
 
 const CONFIG = {
   production: {